LayoutTests/http/tests/cookies/multiple-cookies-expected.txt

This test checks that mulitple cookies are correctly set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check setting several cookies without clearing.
PASS cookie is 'test=foobar'.
PASS cookie is 'test=foobar; test2=foobar'.
PASS cookie is 'test=foobar; test2=foobar; test3=foobar'.
PASS successfullyParsed is true

TEST COMPLETE
